dpwh_agent/requirements.txt
python-dotenv==1.1.0
google-genai>=1.21.1

# Core data stack
pandas==2.2.3
pyarrow>=21.0.0
rapidfuzz==3.9.7